==> scripts/1_importing_fcs_metadata_guide.Rmd <==
---
title: "importing_fcs_metadata"
author: "Hisham M Shaikh"
date: "2023-08-11"
output: html_document

---

**AIM: To extract and wrangle virus counts from viral production assay**\
\
Steps include:\
1. Defining general variables.\
2. Setting up the environment.\
3. Importing and processing metadata file.\
4. Importing FCS files.\
4. Creating a reference FCS file.\
5. Perform a small test.\
6. Extract cytoplots.\
7. Extract counts.\
8. Adjust TE and tidy data.\
\
\
**1. Define general variables**\
Define the working directory `work_dir`.\
This could be done by using `getwd()` function.\

```{r}
work_dir<- normalizePath(paste0(getwd(), "/../")) #adjust working directory accordingly. This is the R project folder
source(here::here("scripts", "1_importing_fcs_metadata_source.R"))
```

Define `project_title`\

```{r}
project_title<- "PE_Cruises_test"
```

To begin with you'll have to define a few variables.\

**2. Setting up environment**\
This is to create a space for your data, metadata, raw data, and ref_fcs file, 
along with a directory for output.\
Also, to install and load all the required libraries.\
Use the function `set_up_vp_count` for this.\

```{r}
source(here::here("scripts", "1_importing_fcs_metadata_source.R"))
set_up_vp_count()
```

**3.  Import and process metadata file**\
The function `metadata_processing` accepts file path to the metadata file. It 
also takes care of the dates.
Please ensure that you have the following parameters mentioned in your metadata file.\
ADD THE PARAMETERS HERE\

```{r}
metadata_processing(paste0(work_dir,"Metadata/PE_Cruises_VP_Metadata.xlsx"), extension = ".xlsx", project_title = "PE_Cruises_test")

metadata <- metadata[1:10,]
```
This function also moves a copy of formatted metadata file in `./data/metadata` folder
and also in the global environment\
\
**4. Importing FCS files**\
We can utilize the `Sample_Name` variable in the `metadata` file to copy files for analyses.\
We need to provide the origin of these FCS files. Files will be copied to `./data/raw_data`.\

```{r}
origin<- paste0(work_dir,"./Data/") # Use absolute path

import_fcs(origin, project_title)
```


**5. Creating Ref file**\ #NOT NEEDED ANYMORE
You need to provide an fcs file. Preferably a file with viruses, as they have more 
events. This is important when analysing viruses and bacteria from different staining protocols.\
```{r}
#ref_fcs_create("vi201104.052")
```

**6. Setting Up for running flowcore**\
Add a gating function .\
```{r}
#This is already in the source script, but if you want to adjust, you can do it here.
{
  polycut<- matrix(c(1.1, 2.0, 2.5, 3.2, 3.7, 3.7, 2.0, 0.6,
                     1.9, 1.5, 1.5, 1.8, 2.8, 3.7, 3.2, 2.75), nrow = 8, ncol=2)
  colnames(polycut)<- c("SSC-H", "FL1-H")
  bgate<- polygonGate(.gate = polycut, filterId = "Bacteria" )
  vgate<- rectangleGate(filterId= "Viruses", "SSC-H" = c(-0.1,1.35), "FL1-H" = c(-0.1, 1.7)) 
  #Different bacterial gates for viral vs bacterial stained samples. We need different gates cause sometimes the bacterial
  #and viral samples have a slight shift.
  HNA_Bacteria_b<- rectangleGate(filterId="HNA_Bacteria",
                                 "SSC-H"=c(0.6, 3.5), "FL1-H"=c(2.15, 3.5))
  LNA_Bacteria_b<- rectangleGate(filterId="LNA_Bacteria",
                                 "SSC-H"=c(1.0, 3.5), "FL1-H"=c(1.5, 2.15))
  
  HNA_Bacteria_v<- rectangleGate(filterId="HNA_Bacteria",
                                 "SSC-H"=c(0.6, 3.5), "FL1-H"=c(2.3, 3.5))
  LNA_Bacteria_v<- rectangleGate(filterId="LNA_Bacteria",
                                 "SSC-H"=c(1.0, 3.5), "FL1-H"=c(1.5, 2.3))
  
  #in case it doesn't, we could define the gate as following
  HNA_Bacteria_bv<- rectangleGate(filterId="HNA_Bacteria",
                                  "SSC-H"=c(0.6, 3.7), "FL1-H"=c(2.3, 3.6))
  LNA_Bacteria_bv<- rectangleGate(filterId="LNA_Bacteria",
                                  "SSC-H"=c(1.0, 3.7), "FL1-H"=c(1.5, 2.3))
  
  #same viral gates as we don't utilise the viral info from bacterial samples
  v1<- rectangleGate(filterId="V1", 
                     "SSC-H"= c(-0.1, 0.90), "FL1-H"= c(-0.1, 0.8)) 
  v2<- rectangleGate(filterId="V2", 
                     "SSC-H"= c(-0.1, 0.90), "FL1-H"= c(0.8, 1.25)) 
  v3<- rectangleGate(filterId="V3", 
                     "SSC-H"= c(-0.1, 1.3), "FL1-H"= c(1.25, 1.7))
  
  detectors<- c("FSC-H", "SSC-H", "FL1-H", "FL2-H", "FL3-H")
  
  translist_bv<- transformList(detectors, logTransform())
}

populate_gate_df()
```

**7.Perform a small test**\
This is to check if the get counts and get plots work allright.\
If the gates look off then you'll have to provide the values of gates individually. \

```{r}
#small test
get_bv_stats(test = T)
get_bv_plots(test = T)

```
Looking at these, you can decide to update your gates. \
If the tests are alright, then run extract cytoplot.\
If that looks good, you can extract counts from them.\

```{r}
get_bv_plots()
get_bv_stats()
```


**8. Combine Counts and Metadata**\

```{r}
counts <- as.data.frame(read_csv("./results/NJ2020_counts.csv"))
combine_metadata_counts()
```

**9.0 TE Examination**\

```{r}
#Separate the dataframe containing TE
TE<- counts_metadata[counts_metadata$Sample_Type == 'TE',]
plotly::ggplotly(ggplot(data = TE[TE$Staining_Protocol == 'Viruses',], aes(x = Sample_Name ,y = V1V2V3))+

                                      geom_point(shape = 1))
#Use this to identify off TEs. Perhaps, we could get rid of all the TEs that are above 2000. Ideally TEs should eb below 500, but the FCM was acting up.
TE<- TE%>% dplyr::filter(V1V2V3 <2000)
plotly::ggplotly(ggplot(data = TE[TE$Staining_Protocol == 'Viruses',], aes(x = Sample_Name ,y = V1V2V3))+
                                    geom_point(shape = 1))
#Looking at this i could also get rid of values above 1000 as there aren't any consecutive TEs above 1000
TE<- TE %>% dplyr::filter( V1V2V3 <1000)
plotly::ggplotly(ggplot(data = TE[TE$Staining_Protocol == 'Viruses',], aes(x = Sample_Name ,y = V1V2V3))+
                                    geom_point(shape = 1))

#Now we need to remove all the TEs we got rid of

calc_TE()


#Let's examine the output of this script.
plotly::ggplotly(ggplot(data = counts_metadata, aes(x = Sample_Name ,y = TE_Vi))+
                                    geom_point(shape = 1))
#The TE values look good.Except for the 800 something values
#We can now adjust the TE values

adjust_TE()

```
**10. Visualization**\
```{r}

NJ2020<- read.csv("./results/NJ2020.csv")

viral_count_overview_plots(NJ2020)
bacterial_count_overview_plots(NJ2020)
```
